Auto-append @Recently-Snapshot, and fix mountpoints
The FileSet is expecting absolute paths for the Exclude -> File
directive, not relative ones. To make this easier, automatically derive
the absolute paths (as well as the inclusions) from the list of SMB
mount points, which is a separate variable. Remove the standalone
vaulted variable for a fileset.

Relatedly, fix the mountpoints for some SMB shares. I accidentally
mounted a few of them underneath my user's home directory.