Move the account keys to group_vars/all
There isn't anything to be gained from a dedicated play against the
control node (localhost). I thought it was possible to batch multiple
requests and to correlate them all to a single TXT record (DNS-01)
validation, but that's not how it works at all.

Note that many of the tasks will still be done from the control node,
but that's just as a matter of cleanliness. The important thing is that
there won't be a sequential set of tasks run against it.